+++ src/components/buttons/InfoButton.tsx
@@ -1,4 +1,4 @@
-import { InfoIcon } from "../icons/InfoIcon";
+import { InfoIcon } from "../../icons/InfoIcon";
 import { withBaseButton } from "./BaseButton";
 
 const InfoButton = withBaseButton(InfoIcon);
