+++ src/components/buttons/ReloadButton.tsx
@@ -1,4 +1,4 @@
-import { ReloadIcon } from "../icons/ReloadIcon";
+import { ReloadIcon } from "../../icons/ReloadIcon";
 import { withBaseButton } from "./BaseButton";
 
 const ReloadButton = withBaseButton(ReloadIcon);
