+++ src/components/buttons/ThemeButton.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
-import { Cloudy } from "../icons/Cloudy";
-import { Night } from "../icons/Night";
-import { Sunny } from "../icons/Sunny";
+import { Cloudy } from "../../icons/Cloudy";
+import { Night } from "../../icons/Night";
+import { Sunny } from "../../icons/Sunny";
 
 interface IProps {
   theme: number;
