+++ src/components/buttons/__tests__/BaseButton.test.tsx
@@ -1,6 +1,6 @@
 import { shallow } from "enzyme";
 import * as React from "react";
-import { InfoIcon } from "../../icons/InfoIcon";
+import { InfoIcon } from "../../../icons/InfoIcon";
 import { withBaseButton } from "../BaseButton";
 
 describe("withBaseButton", () => {
